fix(component): relocate types
